requirements.txt
# keras
# h5py

# tensorflow
# tensorflow-gpu
opencv-python
pathos

pillow
pandas
matplotlib
# cython
jupyter
